Move postgres preamble to lib/
